Reworked navigation bars, now they are different for different pages. App component is aware of the current page and sets it to the right navigation bar. Added auth checker to the navigation bar and landing to check for the user login.
Added client table to display all users/clients inside the /users collection in firestore. Added actions like add a user and delete a user to the table.

Reworked login and contact modals. Gave modals a separate folder. Added a service to reload the table when another user is added or a user is deleted.

added a button-default class which removes the default styles from the button.